[ops] - adding an emails file for scalability and removing the sendEmail function from the ticketController to allow for resuability.
